Refactoring of imports of sub-packages
