+++ src/mlpro/bf/streams/__init__.py
@@ -1,2 +1,1 @@
 from mlpro.bf.streams.basics import *
-from mlpro.bf.streams.streams import *
