+++ src/mlpro/oa/streams/tasks/__init__.py
@@ -1,4 +1,4 @@
-from mlpro.oa.streams.tasks.boundarydetectors import *
+from mlpro.oa.streams.tasks.boundarydetector import BoundaryDetector
 from mlpro.oa.streams.tasks.normalizers import *
 from mlpro.oa.streams.tasks.clusteranalyzers import *
 from mlpro.oa.streams.tasks.anomalydetectors import *
