+++ src/mlpro/oa/streams/tasks/anomalydetectors/__init__.py
@@ -1,4 +1,4 @@
+from mlpro.oa.streams.tasks.anomalydetectors.basics import AnomalyDetector
 from mlpro.oa.streams.tasks.anomalydetectors.anomalies import *
-from mlpro.oa.streams.tasks.anomalydetectors.basics import *
+from mlpro.oa.streams.tasks.anomalydetectors.paga_detectors import AnomalyDetectorPAGA
 from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors import *
-from mlpro.oa.streams.tasks.anomalydetectors.paga_detectors import *
