+++ src/mlpro/oa/streams/tasks/anomalydetectors/anomalies/__init__.py
@@ -1,6 +1,6 @@
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.basics import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.point import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.group import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.drift import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.contextual import *
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.basics import Anomaly
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.point import PointAnomaly
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.group import GroupAnomaly
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.drift import DriftAnomaly
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.contextual import ContextualAnomaly
 from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased import *
