+++ src/mlpro/oa/streams/tasks/anomalydetectors/anomalies/clusterbased/__init__.py
@@ -1,7 +1,7 @@
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.basics import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.density import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.disappearance import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.drift import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.enlargement import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.shrinkage import *
-from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.new_cluster import *
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.basics import CBAnomaly
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.density import ClusterDensityVariation
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.disappearance import ClusterDisappearance
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.drift import ClusterDrift
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.enlargement import ClusterEnlargement
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.shrinkage import ClusterShrinkage
+from mlpro.oa.streams.tasks.anomalydetectors.anomalies.clusterbased.new_cluster import NewClusterAppearance
