+++ src/mlpro/oa/streams/tasks/anomalydetectors/cb_detectors/__init__.py
@@ -1,7 +1,7 @@
-from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.basics import *
-from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.disappearance_detector import *
-from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.drift_detector import *
-from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.density_change_detector import *
-from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.geo_size_change_detector import *
-from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.new_cluster_detector import *
-from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.size_change_detector import *
+from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.basics import AnomalyDetectorCB
+from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.disappearance_detector import ClusterDisappearanceDetector
+from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.drift_detector import ClusterDriftDetector
+from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.density_change_detector import ClusterDensityChangeDetector
+from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.geo_size_change_detector import ClusterGeometricSizeChangeDetector
+from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.new_cluster_detector import NewClusterDetector
+from mlpro.oa.streams.tasks.anomalydetectors.cb_detectors.size_change_detector import ClusterSizeChangeDetector
