--- src/mlpro/oa/streams/tasks/classifiers/__init__.py
@@ -1,2 +1,0 @@
-from mlpro.oa.streams.tasks.clusteranalyzers.clusters import *
-from mlpro.oa.streams.tasks.clusteranalyzers.basics import *
