--- src/mlpro/oa/streams/tasks/classifiers/classes/__init__.py
@@ -1,2 +1,0 @@
-from mlpro.oa.streams.tasks.clusteranalyzers.clusters.basics import *
-from mlpro.oa.streams.tasks.clusteranalyzers.clusters.centroid import *
