+++ src/mlpro/oa/streams/tasks/clusteranalyzers/__init__.py
@@ -1,1 +1,2 @@
 from mlpro.oa.streams.tasks.clusteranalyzers.basics import ClusterAnalyzer
+from mlpro.oa.streams.tasks.clusteranalyzers.clusters import *
