+++ src/mlpro/oa/streams/tasks/clusteranalyzers/clusters/properties/__init__.py
@@ -1,2 +1,2 @@
 from mlpro.oa.streams.tasks.clusteranalyzers.clusters.properties.basics import *
-from mlpro.oa.streams.tasks.clusteranalyzers.clusters.properties.centroid import *
+from mlpro.oa.streams.tasks.clusteranalyzers.clusters.properties.centroid import Centroid, cprop_center_geo, cprop_center_geo1, cprop_center_geo2, cprop_centroid, cprop_centroid1, cprop_centroid2, cprop_centroid_prev, cprop_centroid_prev1, cprop_centroid_prev2
