+++ test/howtos/oa/howto_oa_streams_pp_102_rearranger_window_bd_normminmax_3d.py
@@ -33,11 +33,17 @@ You will learn:
 
 """
 
+from datetime import datetime
 
-from mlpro.bf.streams import *
-from mlpro.bf.streams.streams import *
+from mlpro.bf.various import Log
+from mlpro.bf.mt import Task
+from mlpro.bf.ops import Mode
+from mlpro.bf.plot import PlotSettings
+
+from mlpro.bf.streams.streams import StreamProviderMLPro, StreamMLProRnd10D
 from mlpro.bf.streams.tasks import RingBuffer, Rearranger
-from mlpro.oa.streams import *
+from mlpro.oa.streams import OAWorkflow, OAScenario
+from mlpro.oa.streams.tasks import BoundaryDetector, NormalizerMinMax
 
 
 
@@ -139,7 +145,7 @@ else:
     cycle_limit = 60
     logging     = Log.C_LOG_NOTHING
     visualize   = False
-    step_rate   = 5
+    step_rate   = 1
 
 
 # 2 Instantiate the stream scenario
