+++ test/howtos/oa/howto_oa_streams_pp_103_rearranger_window_bd_normminmax_nd.py
@@ -11,10 +11,11 @@
 ## -- 2023-01-09  1.1.0     DA       User input of cycles and visualization step rate
 ## -- 2023-04-10  1.2.0     DA       Refactoring after changes on class OAScenario
 ## -- 2023-05-20  1.2.1     DA       Registered handler of boundary detector to window
+## -- 2024-10-29  1.2.2     DA       Minor corrections
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.2.1 (2023-05-20)
+Ver. 1.2.2 (2024-10-29)
 
 This module is an example of adaptive normalization of streaming data using MinMax normalizer. To 
 this regard, an online-adadptive custom scenario is set up. It combines a native 10-dimensional 
@@ -96,7 +97,7 @@ class MyAdaptiveScenario (OAScenario):
         # 2.2.3 Boundary detector 
         task_bd = BoundaryDetector( p_name='T3 - Boundary Detector', 
                                     p_ada=True, 
-                                    p_visualize=True,   
+                                    p_visualize=p_visualize,   
                                     p_logging=p_logging )
 
         task_window.register_event_handler( p_event_id=RingBuffer.C_EVENT_DATA_REMOVED, p_event_handler=task_bd.adapt_on_event )
@@ -137,9 +138,10 @@ if __name__ == "__main__":
 
 else:
     # 1.2 Parameters for internal unit test
-    cycle_limit = 2
-    logging = Log.C_LOG_NOTHING
-    visualize = False
+    cycle_limit = 60
+    logging     = Log.C_LOG_NOTHING
+    visualize   = False
+    step_rate   = 1 
 
 
 # 2 Instantiate the stream scenario
