+++ test/howtos/oa/howto_oa_streams_pp_104_rearranger_window_bd_normminmax_2d_3d_nD.py
@@ -34,6 +34,7 @@ from mlpro.bf.streams import *
 from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.tasks import RingBuffer, Rearranger
 from mlpro.oa.streams import *
+from mlpro.oa.streams.tasks import BoundaryDetector, NormalizerMinMax
