+++ test/howtos/oa/howto_oa_streams_pp_121_complex_preprocessing.py
@@ -43,7 +43,7 @@ from mlpro.bf.streams.streams import *
 from mlpro.bf.streams.tasks import *
 
 from mlpro.oa.streams import *
-from mlpro.oa.streams.tasks import BoundaryDetector, NormalizerMinMax
+from mlpro.oa.streams.tasks import BoundaryDetector, NormalizerMinMax, NormalizerZTransform
